feat(groups): implemented default group handling
implemented add user to default groups on creation or activation. the
user now will be removed on deactivating the account. based on ACL the
user management permission maybe has no permission to add a user to some
groups, this feature is done over a technical management access which
can be restricted for this functionality on ldap server side.