feat(bugtracker): added configurable bugtracking url
added a config with default to github issues page. this allows to
override the found a bug link on the footer to an internal bugtracking
tool. please keep in mind, that found bugs also reported to public
github, so everybody can participate from this finding. also added
config value on maintenance view, which helps to prioritize the account
maintenance.